function moved from PipelineBuilder  to BuildPipelineConfigService and DeploymentPipelineConfigService
