+++ pkg/pipeline/BuildPipelineConfigService.go
@@ -22,6 +22,7 @@ import (
 	"errors"
 	"fmt"
 	bean2 "github.com/devtron-labs/devtron/api/bean"
+	"github.com/devtron-labs/devtron/client/argocdServer/application"
 	"github.com/devtron-labs/devtron/internal/sql/repository"
 	app2 "github.com/devtron-labs/devtron/internal/sql/repository/app"
 	"github.com/devtron-labs/devtron/internal/sql/repository/appWorkflow"
@@ -2310,4 +2311,162 @@ func (impl *PipelineBuilderImpl) buildResponses() []bean.ResponseSchemaObject {
 	responseSchemaObjects = append(responseSchemaObjects, response400)
 	responseSchemaObjects = append(responseSchemaObjects, response401)
 	return responseSchemaObjects
+}
+
+func (impl *PipelineBuilderImpl) BuildArtifactsForParentStage(cdPipelineId int, parentId int, parentType bean2.WorkflowType, ciArtifacts []bean.CiArtifactBean, artifactMap map[int]int, limit int, parentCdId int) ([]bean.CiArtifactBean, error) {
+	var ciArtifactsFinal []bean.CiArtifactBean
+	var err error
+	if parentType == bean2.CI_WORKFLOW_TYPE {
+		ciArtifactsFinal, err = impl.BuildArtifactsForCIParent(cdPipelineId, parentId, parentType, ciArtifacts, artifactMap, limit)
+	} else if parentType == bean2.WEBHOOK_WORKFLOW_TYPE {
+		ciArtifactsFinal, err = impl.BuildArtifactsForCIParent(cdPipelineId, parentId, parentType, ciArtifacts, artifactMap, limit)
+	} else {
+		//parent type is PRE, POST or DEPLOY type
+		ciArtifactsFinal, _, _, _, err = impl.BuildArtifactsForCdStage(parentId, parentType, ciArtifacts, artifactMap, true, limit, parentCdId)
+	}
+	return ciArtifactsFinal, err
+}
+
+func (impl *PipelineBuilderImpl) BuildArtifactsForCdStage(pipelineId int, stageType bean2.WorkflowType, ciArtifacts []bean.CiArtifactBean, artifactMap map[int]int, parent bool, limit int, parentCdId int) ([]bean.CiArtifactBean, map[int]int, int, string, error) {
+	//getting running artifact id for parent cd
+	parentCdRunningArtifactId := 0
+	if parentCdId > 0 && parent {
+		parentCdWfrList, err := impl.cdWorkflowRepository.FindArtifactByPipelineIdAndRunnerType(parentCdId, bean2.CD_WORKFLOW_TYPE_DEPLOY, 1)
+		if err != nil || len(parentCdWfrList) == 0 {
+			impl.logger.Errorw("error in getting artifact for parent cd", "parentCdPipelineId", parentCdId)
+			return ciArtifacts, artifactMap, 0, "", err
+		}
+		parentCdRunningArtifactId = parentCdWfrList[0].CdWorkflow.CiArtifact.Id
+	}
+	//getting wfr for parent and updating artifacts
+	parentWfrList, err := impl.cdWorkflowRepository.FindArtifactByPipelineIdAndRunnerType(pipelineId, stageType, limit)
+	if err != nil {
+		impl.logger.Errorw("error in getting artifact for deployed items", "cdPipelineId", pipelineId)
+		return ciArtifacts, artifactMap, 0, "", err
+	}
+	deploymentArtifactId := 0
+	deploymentArtifactStatus := ""
+	for index, wfr := range parentWfrList {
+		if !parent && index == 0 {
+			deploymentArtifactId = wfr.CdWorkflow.CiArtifact.Id
+			deploymentArtifactStatus = wfr.Status
+		}
+		if wfr.Status == application.Healthy || wfr.Status == application.SUCCEEDED {
+			lastSuccessfulTriggerOnParent := parent && index == 0
+			latest := !parent && index == 0
+			runningOnParentCd := parentCdRunningArtifactId == wfr.CdWorkflow.CiArtifact.Id
+			if ciArtifactIndex, ok := artifactMap[wfr.CdWorkflow.CiArtifact.Id]; !ok {
+				//entry not present, creating new entry
+				mInfo, err := parseMaterialInfo([]byte(wfr.CdWorkflow.CiArtifact.MaterialInfo), wfr.CdWorkflow.CiArtifact.DataSource)
+				if err != nil {
+					mInfo = []byte("[]")
+					impl.logger.Errorw("Error in parsing artifact material info", "err", err)
+				}
+				ciArtifact := bean.CiArtifactBean{
+					Id:                            wfr.CdWorkflow.CiArtifact.Id,
+					Image:                         wfr.CdWorkflow.CiArtifact.Image,
+					ImageDigest:                   wfr.CdWorkflow.CiArtifact.ImageDigest,
+					MaterialInfo:                  mInfo,
+					LastSuccessfulTriggerOnParent: lastSuccessfulTriggerOnParent,
+					Latest:                        latest,
+					Scanned:                       wfr.CdWorkflow.CiArtifact.Scanned,
+					ScanEnabled:                   wfr.CdWorkflow.CiArtifact.ScanEnabled,
+				}
+				if !parent {
+					ciArtifact.Deployed = true
+					ciArtifact.DeployedTime = formatDate(wfr.StartedOn, bean.LayoutRFC3339)
+				}
+				if runningOnParentCd {
+					ciArtifact.RunningOnParentCd = runningOnParentCd
+				}
+				ciArtifacts = append(ciArtifacts, ciArtifact)
+				//storing index of ci artifact for using when updating old entry
+				artifactMap[wfr.CdWorkflow.CiArtifact.Id] = len(ciArtifacts) - 1
+			} else {
+				//entry already present, updating running on parent
+				if parent {
+					ciArtifacts[ciArtifactIndex].LastSuccessfulTriggerOnParent = lastSuccessfulTriggerOnParent
+				}
+				if runningOnParentCd {
+					ciArtifacts[ciArtifactIndex].RunningOnParentCd = runningOnParentCd
+				}
+			}
+		}
+	}
+	return ciArtifacts, artifactMap, deploymentArtifactId, deploymentArtifactStatus, nil
+}
+
+func (impl *PipelineBuilderImpl) BuildArtifactsForCIParent(cdPipelineId int, parentId int, parentType bean2.WorkflowType, ciArtifacts []bean.CiArtifactBean, artifactMap map[int]int, limit int) ([]bean.CiArtifactBean, error) {
+	artifacts, err := impl.ciArtifactRepository.GetArtifactsByCDPipeline(cdPipelineId, limit, parentId, parentType)
+	if err != nil {
+		impl.logger.Errorw("error in getting artifacts for ci", "err", err)
+		return ciArtifacts, err
+	}
+	for _, artifact := range artifacts {
+		if _, ok := artifactMap[artifact.Id]; !ok {
+			mInfo, err := parseMaterialInfo([]byte(artifact.MaterialInfo), artifact.DataSource)
+			if err != nil {
+				mInfo = []byte("[]")
+				impl.logger.Errorw("Error in parsing artifact material info", "err", err, "artifact", artifact)
+			}
+			ciArtifacts = append(ciArtifacts, bean.CiArtifactBean{
+				Id:           artifact.Id,
+				Image:        artifact.Image,
+				ImageDigest:  artifact.ImageDigest,
+				MaterialInfo: mInfo,
+				ScanEnabled:  artifact.ScanEnabled,
+				Scanned:      artifact.Scanned,
+			})
+		}
+	}
+	return ciArtifacts, nil
+}
+
+func parseMaterialInfo(materialInfo json.RawMessage, source string) (json.RawMessage, error) {
+	if source != "GOCD" && source != "CI-RUNNER" && source != "EXTERNAL" {
+		return nil, fmt.Errorf("datasource: %s not supported", source)
+	}
+	var ciMaterials []repository.CiMaterialInfo
+	err := json.Unmarshal(materialInfo, &ciMaterials)
+	if err != nil {
+		println("material info", materialInfo)
+		println("unmarshal error for material info", "err", err)
+	}
+	var scmMapList []map[string]string
+
+	for _, material := range ciMaterials {
+		scmMap := map[string]string{}
+		var url string
+		if material.Material.Type == "git" {
+			url = material.Material.GitConfiguration.URL
+		} else if material.Material.Type == "scm" {
+			url = material.Material.ScmConfiguration.URL
+		} else {
+			return nil, fmt.Errorf("unknown material type:%s ", material.Material.Type)
+		}
+		if material.Modifications != nil && len(material.Modifications) > 0 {
+			_modification := material.Modifications[0]
+
+			revision := _modification.Revision
+			url = strings.TrimSpace(url)
+
+			_webhookDataStr := ""
+			_webhookDataByteArr, err := json.Marshal(_modification.WebhookData)
+			if err == nil {
+				_webhookDataStr = string(_webhookDataByteArr)
+			}
+
+			scmMap["url"] = url
+			scmMap["revision"] = revision
+			scmMap["modifiedTime"] = _modification.ModifiedTime
+			scmMap["author"] = _modification.Author
+			scmMap["message"] = _modification.Message
+			scmMap["tag"] = _modification.Tag
+			scmMap["webhookData"] = _webhookDataStr
+			scmMap["branch"] = _modification.Branch
+		}
+		scmMapList = append(scmMapList, scmMap)
+	}
+	mInfo, err := json.Marshal(scmMapList)
+	return mInfo, err
 }
